+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/model/BalanceTest.kt
@@ -18,18 +18,18 @@ class BalanceTest {
 
     private var amount: BigDecimal = 0.0.bd
     private var subUnitToUnit: BigDecimal = 100_000.0.bd
-    private lateinit var mintedToken: MintedToken
+    private lateinit var token: Token
 
     @Before
     fun setup() {
-        mintedToken = MintedToken("omg:1234", "OMG", "OmiseGO", subUnitToUnit)
+        token = Token("omg:1234", "OMG", "OmiseGO", subUnitToUnit)
     }
 
     @Test
     fun `test very big amount and small subUnitToUnit`() {
         // 36 zeros
         amount = 1_000_000_000_000_000_000_000_000_000_000_000_000.0.bd
-        val token = mintedToken.copy(subunitToUnit = 1.0.bd)
+        val token = token.copy(subunitToUnit = 1.0.bd)
         val balance = Balance(amount, token)
         balance.displayAmount() shouldEqual "1,000,000,000,000,000,000,000,000,000,000,000,000.00"
     }
@@ -37,7 +37,7 @@ class BalanceTest {
     @Test
     fun `test small amount and very big subUnitToUnit `() {
         amount = 1.0.bd
-        val token = mintedToken.copy(subunitToUnit = 1_000_000_000_000_000_000_000_000_000_000_000_000.0.bd)
+        val token = token.copy(subunitToUnit = 1_000_000_000_000_000_000_000_000_000_000_000_000.0.bd)
         val balance = Balance(amount, token)
 
         // 36 decimal
@@ -47,7 +47,7 @@ class BalanceTest {
     @Test
     fun `test zero display`() {
         amount = 0.0.bd
-        val token = mintedToken.copy(subunitToUnit = 1.0.bd)
+        val token = token.copy(subunitToUnit = 1.0.bd)
         val balance = Balance(amount, token)
 
         balance.displayAmount(0) shouldEqual "0"
@@ -56,7 +56,7 @@ class BalanceTest {
     @Test
     fun `test zero display with 5 precision`() {
         amount = 0.0.bd
-        val token = mintedToken.copy(subunitToUnit = 1.0.bd)
+        val token = token.copy(subunitToUnit = 1.0.bd)
         val balance = Balance(amount, token)
 
         balance.displayAmount(5) shouldEqual "0.00000"
@@ -65,7 +65,7 @@ class BalanceTest {
     @Test
     fun `test division with cool divisor number and display very large precision correctly`() {
         val coolDivisorNumber = 998_001.0.bd
-        val token = mintedToken.copy(subunitToUnit = coolDivisorNumber)
+        val token = token.copy(subunitToUnit = coolDivisorNumber)
         val balance = Balance(1.0.bd, token)
 
         balance.displayAmount(99) shouldEqual "0.000001002003004005006007008009010011012013014015016017018019020021022023024025026027028029030031032"
@@ -73,8 +73,8 @@ class BalanceTest {
 
     @Test
     fun `test balance + another balance should be correct`() {
-        val balance1 = Balance(1_999_000_000_000.0.bd, MintedToken("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
-        val balance2 = Balance(9_999_000_000_000.0.bd, MintedToken("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
+        val balance1 = Balance(1_999_000_000_000.0.bd, Token("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
+        val balance2 = Balance(9_999_000_000_000.0.bd, Token("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
 
         val result = balance1 + balance2
         result.amount shouldEqual 11_998_000_000_000.0.bd
@@ -82,8 +82,8 @@ class BalanceTest {
 
     @Test
     fun `test balance - another balance should be correct`() {
-        val balance1 = Balance(11_998_000_000_000.0.bd, MintedToken("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
-        val balance2 = Balance(9_999_000_000_000.0.bd, MintedToken("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
+        val balance1 = Balance(11_998_000_000_000.0.bd, Token("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
+        val balance2 = Balance(9_999_000_000_000.0.bd, Token("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
 
         val result = balance1 - balance2
         result.amount shouldEqual 1_999_000_000_000.0.bd
@@ -94,8 +94,8 @@ class BalanceTest {
 
     @Test(expected = UnsupportedOperationException::class)
     fun `test balance + incompatible balance should throw UnSupportOperationException`() {
-        val balance1 = Balance(1_999_000_000_000.0.bd, MintedToken("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
-        val balance2 = Balance(9_999_000_000_000.0.bd, MintedToken("ETH:8bcda572-9411-43c8-baae-cd56eb0155f3", "ETH", "Etherium", 10000.0.bd))
+        val balance1 = Balance(1_999_000_000_000.0.bd, Token("OMG:8bcda572-9411-43c8-baae-cd56eb0155f3", "OMG", "OmiseGO", 10000.0.bd))
+        val balance2 = Balance(9_999_000_000_000.0.bd, Token("ETH:8bcda572-9411-43c8-baae-cd56eb0155f3", "ETH", "Etherium", 10000.0.bd))
 
         balance1 + balance2
     }